Dodanie nowej funkcjonalnosci do programu

+++ src/main/java/com/project/pp/DTO/ResultDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @AllArgsConstructor
 public class ResultDTO {
 
-    private long resultTime;
+    private double resultTime;
 
 
 }
